feat: burner wallets deployer contract
The `BurnerWalletDeployer` contract allows to deploy OZ accounts to be used as burner wallets.

This is done with the `deploy_burner_wallet` which takes the account public key as parameter and returns
the created account address.

To not deploy a same account forever, this contract uses an internal `salt` starting from 0 and increased
of 1 after each new burner wallet.